+++ backend/src/main/java/ru/laspace/backend/service/Pr04BuilderService.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface Pr04BuilderService {
-    String build(Long numRp, Long numKa, Integer rnf);
+    String build(Integer numRp, Integer numKa, Integer rnf);
 }
