+++ backend/src/main/java/ru/laspace/backend/service/Pr04Service.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface Pr04Service {
-    String generateAndSave(Long numRp, Long numKa);
+    String generateAndSave(Integer numRp, Integer numKa);
 }
