+++ backend/src/main/java/ru/laspace/backend/service/ProgramsService.java
@@ -3,5 +3,5 @@ package ru.laspace.backend.service;
 import ru.laspace.backend.dto.programs.ProgramCreateRequest;
 
 public interface ProgramsService {
-    Long saveProgram(ProgramCreateRequest request);
+    Integer saveProgram(ProgramCreateRequest request);
 }
