+++ backend/src/main/java/ru/laspace/backend/service/Vp01BuilderService.java
@@ -1,5 +1,5 @@
 package ru.laspace.backend.service;
 
 public interface Vp01BuilderService {
-    String build(Long numRp, Long numKa, Integer rnf);
+    String build(Integer numRp, Integer numKa, Integer rnf);
 }
