+++ backend/src/main/java/ru/laspace/backend/service/impl/Pr03BuilderServiceImpl.java
@@ -28,7 +28,7 @@ public class Pr03BuilderServiceImpl implements Pr03BuilderService {
     private final ProgramsOnaRepository programsOnaRepository;
 
     @Override
-    public String build(Long numRp, Long numKa, Integer rnf) {
+    public String build(Integer numRp, Integer numKa, Integer rnf) {
         ProgramsMain main = programsMainRepository.findByNumRpAndNumKa(numRp, numKa)
                 .orElseThrow(() -> new RuntimeException(
                         "ПРЦА не найдена: numRp=" + numRp + ", numKa=" + numKa));
