+++ backend/src/main/java/ru/laspace/backend/service/impl/Pr04BuilderServiceImpl.java
@@ -28,7 +28,7 @@ public class Pr04BuilderServiceImpl implements Pr04BuilderService {
     private final ProgramsOnaRepository programsOnaRepository;
 
     @Override
-    public String build(Long numRp, Long numKa, Integer rnf) {
+    public String build(Integer numRp, Integer numKa, Integer rnf) {
         ProgramsMain main = programsMainRepository.findByNumRpAndNumKa(numRp, numKa)
                 .orElseThrow(() -> new RuntimeException(
                         "ПРЦА не найдена: numRp=" + numRp + ", numKa=" + numKa));
